Generalize this approach to all Begin/End pairs with a macro
The macro simply generates a variadic template function with the
specified name.  It would be nice if we could have a semi-colon at the
end of the macro declaration, but the compiler warns about an "extra
semicolon" in that case.

In addition to VecScatterBegin/End, we now support
MatAssemblyBegin/End, VecAssemblyBegin/End, and
VecGhostUpdateBegin/End.