Remove potential deadlock with Singletons and RemoteElem
We had a potential deadlock before as reported by helgrind. We would
acquire the `remote_elem_mtx` first during setup and then
`singleton_mtx`. However during destruction we acquire `singleton_mtx`
first and then `remote_elem_mtx`. This has the potential to deadlock. So
why not let the singleton code handle all the locking? Seems like
singleton code should do that kind of thing and we should leverage it in
`RemoteElem` if we are going to go through the process of inheriting
from `Singleton`. This reports helgrind clean now