Turn off gdb backtraces by default
The user will now have to opt-in to gdb backtraces instead of opting out
with `--without-gdb-command`. There are many times that I forget to
configure with `--without-gdb-command` and then I sadly wait for a
minute to get my prompt back or switch to another terminal to kill the
process. For all of MOOSE testing we run with `--no-gdb-backtrace`. In
general if I need a backtrace I will compile in debug mode and print a
backtrace. I almost never make use of the automatic backtrace output;
instead I just dislike it. There was some general support for this
change on our libmesh slack channel