PointLocatorTree: avoid Elem::contains_point() warning when using custom tolerances
Now that we are correctly setting a custom contains_point() tolerance on
the PointLocator within MeshFunction (#2849), we are seeing the following
warning from Elem::contains_point():

    WARNING: Resizing bounding box to match user-specified tolerance!
    In the future, calls to Elem::contains_point() with tol > TOLERANCE
    will be more optimized, but should not be used
    to search for points 'close to' elements!
    Instead, use Elem::close_to_point() for this purpose.

The issue is that we should really be testing Elem::close_to_point()
in the PointLocator (and associated classes) when the user passes in a
custom tolerance > TOLERANCE, since that indicates the user's intent
to locate "nearby" Elems rather than "strictly containing" Elems.

This commit updates the relevant contains_point() checks in
PointLocatorTree and TreeNode<N>::find_element() so that we should no
longer see this warning. Note that we still call the potentially
optimized (currently only for Tri3 and Tet4) contains_point() routine
if the user does not set a custom PointLocator tolerance.