MeshFunction: add non-const accessor for underlying PointLocator
We already had a const accessor for the PointLocator, so while this is
technically encapsulation-breaking, it isn't something that couldn't
have already been hacked with a const_cast.

This allows us to set flags, etc. directly on the MeshFunction's
underlying PointLocator. Now that the PointLocator can now be either
OctTree-based or Nanoflann-based, we may want to have different
behaviors based on the value of the

LIBMESH_ENABLE_NANOFLANN_POINTLOCATOR

define. It currently only makes sense to call certain APIs on some
subclasses, and we also don't want to have to add a MeshFunction API
for every flag we want to set on the PointLocator.