Read/write numeric vector projection/type info
I ran into an issue where I was reading at a ghost dof location with a
vector that should be ghosted, but when reading `System` information
before this PR we add vectors as default `ParallelType PARALLEL` if the
vectors are not already present in the `System` object; the result was a
failed read and a triggered assertion in debugging modes/a segmentation
fault in optimized modes. So here I add the writing of both vector
projection and type so that users are safe to use ghosted vectors
without considering whether their vector has been created at the time we
read xdr `System` information. I know I'll need to add versioning
information because of this change.

Refs idaholab/moose#17399