Don't duplicate determine_calculations() badly
This was breaking my new unit test, where I was doing a get_xyz()
request but not requesting phi/dphi/d2phi, and the code in FEXYZ
(ironically...) hadn't been updated to recognize that as a valid use
case.

I'm not sure why we're doing determine_calculations() here, it seems
redundant to me, but this is a public function so maybe somebody's
calling it from outside our usual code paths?