ExII inquire signature change, anonymyzing
ExodusII is getting a bit more abstract about their inquiry data type,
and complaining if we don't follow suit.

ex_inquiry was defined by v5.09, so this should be a backwards
compatible change.

I don't like having types depending on the exodus headers in our helper
class headers (wasn't this supposed to be a pImpl idiom originally?) so
making this signature change necessitates moving the method outside of
the header; I made it an anonymous function, which should be fine since
it just sees internal use.