Add non-virtual Elem::vertex_average() and update Elem::centroid()
When the Elem has an elevated p_level, this will affect the Order of
the FE that gets reinit()'d for computing the centroid. Rather than do
any non-const hacking of the p_level value, we just work around this
issue by making a copy of the Elem with non-elevated p_level and
return its centroid instead.

This obviously introduces an additional performance hit, but truly
optimized code should not be calling the base class Elem::centroid()
implementation to begin with. This approach also has the benefit of
not requiring const_cast and avoiding potential thread safety issues.