Add Pyramid5 centroid calculation
We are using the formula for computing the centroid of a Pyramid
described here: https://en.wikipedia.org/wiki/Centroid

Note: to call the const version of build_side_ptr() from within this
member function, I had to namespace it with Elem:: since otherwise
virtual dispatch takes over and there is only a non-const version of
the virtual build_side_ptr() in that case. This seems like it could be
improved by making the const version of the function virtual as
well... but there may be a reason why this wasn't done.