Add Pyramid5::true_centroid()
We compute the centroid of the Pyramid by treating it as a
"degenerate" hexahedron in which all four points on the "top" of the
hexahedron are located at the pyramid apex. Since we use a quadrature
rule with interior points, the non-invertibility of the map at the
apex does not seem to cause any issues with the computation. This
approach is much more straightforward/cleaner than trying to deal with
the Pyramid5's rational basis functions to compute the centroid.

Both Pyramid5::true_centroid() and Hex8::true_centroid() are now
implemented in terms of the static Hex8::centroid_from_points()
function. This same approach could also be applied to the Prism6, and
possibly to the higher-order Hex/Prism/Pyramids as well.