FE cases for TET14
This doesn't include the most important cases: the HIERARCHIC code
that'll finally be possible to add at high orders (and thus worthwhile
to add at low orders), and the SIDE_HIERARCHIC code that'll now be
possible.  But let's make sure the new element's basic functionality
works before we go wild with it.