Helper app to extend+verify embedding matrices
You still have to get the child vertex selection right to use this, but
it'll then calculate the non-binary lines of each matrix for you.

Yes, I had been doing this by hand before.  Yes, that might have been a
dumb idea with Tri7.  It was definitely a dumb idea when I started doing
it for Tet14; most of the matrix lines were still obvious permutations
of the same few patterns, but some ... were not.