Turn assert into error in PBCs::neighbor()
This should fix #2958

I don't like putting opt mode tests into anything called on a single
element; that's just asking to bloat kernel runtimes ... but in this
case, when we don't actually have broken PBC ids/displacements, we never
hit the test on a replicated or serialized mesh, and we never hit the
test when calling PBCs::neighbor() on a local element, and at least
within the library itself it looks like we're only calling
PBCs::neighbor from local elements.  So this change ought to be safe for
performance after all.