SIDE_HIERARCHIC normal derivs might not be zero
The normal in master space doesn't map to the normal in physical space
unless the map is just a scaling + rotation + translation; even just
permuting the vertices of a right triangle causes the map to include a
skew.

So, derivative components which are zero in the normal direction on the
master element necessarily become non-zero in the normal direction on
the physical element.  We'd have to switch SIDE_HIERARCHIC (and any
future SIDE_ classes) to use something other than the default H1
transformation if we wanted to change this.

The alternative here is instead:

1. Zero out normal derivative components in our unit tests (done, here;
   the tests will now pass on triangle meshes)
2. Hope most of our users don't have kernels involving derivatives of
   SIDE_ terms (which is reasonable for standard H1 problems)
3. Hope the remaining users are smart enough to zero out non-tangential
   derivative components manually too (which may be reasonable for users
   doing something more sophisticated than standard H1 problems?)