ex2html.sh: update Documentation link in example headers
This is effectively the same change that was made in 4356ab070, which
is not ideal. A better approach would be for the ex2html.sh script to
use the simple_header.html directly, so that changes to the header are
picked up automatically. This would require an update to the
ex2html.sh script, though, which is more effort than I'm willing to
spend on it at the moment.