Add Elem::side_type()
The aids a helper, the ElemSideBuilder, which builds element
sides while minimizing Elem construction. In order to store
a cache of elements for building sides, we need a way to
index into a structure that stores an element for each element
side type. Elem::side_type() allows us to index into said
structure based on the higher dimensional element.