Remove caching from FE<CLOUGH>
Derek did this for HERMITE all the way back in #112, but people don't
push the limits as much with Clough-Tocher, so the corner cases here
never bit me until this week.

We really need to refactor our FE calculations, not just for the extreme
cases like these classes but for all the shareable operations done for
our other FE types too.