Create even less contention
All the readers are waiting on our writer thread to mark our condition
(`_array_is_present`) as ready before they can even attempt to read, so
there is no data race for our writer thread. Hence we can remove the
lock on the mutex the readers are using. And it's much better this way
because one could imagine that our writer hits the `std::unique_lock`
first which would then prevent our reader threads from even getting to
the condition variable `wait`, which is the logical place we want them
to get to while the writer thread is doing its job.

And the second change is unlocking as soon as we're through waiting
because we are through the read-write portion of the program and are
only reading so it's safe to let everyone through at once