Only use new iterators with --disable-deprecated
The point of the new iterators is to prevent people doing unsafe things
in user code, but that's inherently an API change, so we should give
people some time to make changes.

They ought to mostly be easy changes, though.  Hopefully I'm the only
one doing anything as crazy as the custom iterators in
mesh_communication.C and checkpoint_io.C