Added APIs to allow users set the flag: _children_on_boundary
The motivaiton is that: "automatic" way might not work for cases.

For example, if the flag is on on a subset of processor cores, and
then we do reparitioning, and then we might hit trouble because
the flag is off on the other processors

I try to avoid having a global reduction to have everyone on the same page.
In fact, we know when we want to allow children on boundary in MOOSE.
In this case, we should set the flag to make the code more robust