Don't try to directly use feenableexcept
That function isn't available on Mac.

We have a shim function for this we could have switched to instead of
removing it entirely, but Vikram confirms for me that we don't actually
need this in the default example runs and that he can get by with our
'--enable-fpe' command-line option when he's doing custom runs.

On a related note, our command-line options should probably have better
documentation than "code in the middle of a src/base/ .C file that
doesn't get Doxygen'd with a sign in the comments saying 'Beware of the
Leopard'"