set_hdf5_writing(bool) option
We're going to want to turn this off in Moose, so users' new files will
be as portable as possible to other builds (including new gold
regression test files and including our --disable-hdf5 CI recipes).

We'll leave it on by default in libMesh for backwards compatibility;
anybody who went to the trouble of building with HDF5 manually probably
had a reason to want HDF5 output.