Mark Dense(Sub)Matrix methods final
clang-tidy warned about seeing a virtual method (zero()) called in what
might be a base class constructor, and for internal use we want these
functions to be inlinable if they're called from a leaf class, not
virtual calls.

Not marking the whole classes as "final" because we might as well let
people derive from them if they want, just so long as they're not
relying on overriding the virtual methods.