Always add elements/nodes to mesh when unpacking
The packed range algorithms only work if the object is added as
soon as it is unpacked. But push_parallel_packed_range does all
of its unpacking before the user gets a chance to add objects to
the mesh manually. So we need to make sure we do it automatically
in the unpack routines