Remove __SUNPRO_CC specific code
Any compiler old enough to require such guards would likely no longer
be able to compile the library as a whole. In addition, we do not
currently have (and have not had for some time) regular testing of a
compiler which defines __SUNPRO_CC, so it's not even clear whether the
code currently in the guards still compiles.