Quadrature rules for PRISM20/21
The nodal rules are the only interesting ones here.  The nodal rule for
PRISM20 is the only really interesting one here.  I fear my solution
may be in the uncanny valley of "not bad enough that John is motivated
to replace it with something better" but still "not good enough that the
knowledge he could have done better won't irritate him".