Put report_error() back in a mutex
We may no longer need to wrap the ostream operations, but we still need
to wrap the test for whether report_error() is being called recursively.

Trying to just never call report_error() recursively would be doable for
the moment (the asserts inside `on_command_line` tests in `print_trace`
seem to be the only code path that can currently do it), but I don't see
a way to future proof that - we like to do lots of helpful things when
error reporting, and we like to fill the things we do with assertions.