Update to check_dirichlet_bcid_consistency()
Previously check_dirichlet_bcid_consistency() required that all processors had DirichletBoundary objects
defined identically, otherwise one could get a parallel deadlock due to calling comm.max() on a subset
of processors. In this update we do a comm.set_union() of boundary IDs with DirichetBoundary defined on
them so that we can define DirichletBoundary objects differently on different processors. This is desirable
since in general other libMesh mesh data is not required to be identical on different processors, and this
update to check_dirichlet_bcid_consistency() means that DirichletBoundary also no longer needs to be
identical on different processors.