Comment solve_for_unconstrained_dofs limitations
It took me a few hours to realize and verify (on a 1-D problem,
constraining both C0 and discontinuous quadratic elements to be C1) what
was happening here, and I'd rather not forget and repeat that someday.

Galerkin matrices shouldn't give us the same effects, so if we ever want
absolute reproduceability between C0 and discontinuous extraction
meshes, we might want to either solve steady problems or get our initial
conditions from a global projection solve.