Fix NumericVector::compare
The parallel failures from idaholab/moose#23788 are:

```
libMesh terminating:
No index 5 in ghosted vector.
Vector contains [5,5)
And ghost array {4,0}
```

because prior to this commit if the first local index was equal
to the last local index we would go ahead and attempt to index
into the vector with first local index anyway. Now we avoid that
scenario