Allow edges with different orientations to match
The idea here is that a user should be able to specify an edge block
in the Exodus file simply by specifying their desired edge ids,
without knowing the orientation of the edge in the 3D elem it is
attached to. This means that edges which are attached to multiple
Elems will be added to the BoundaryInfo object multiple times.