Add allSecondOrderMixed() unit test
This is similar to the existing allSecondOrderRange() unit test, but
instead of using a separate subdomain to control which elements are
converted to 2nd-order, we now convert all of them, automatically
skipping the first element which is already SECOND-order.

Also avoid using "new" and including elem-specific header files in
this unit test.