Handle not-quite-orphaned DistributedMesh nodes
The MOOSE test case that triggered a `--distributed-mesh` failure here
was working on 15 or fewer processors, on 16 processors it was working
on my own workstation, and it was working 80% of the time (are some
Parmetis versions not deterministic?) on INL servers.

I'll see if I can come up with a more reproduceable unit test, but in
the meantime this at least fixes the issue for me.