Store the number of steps represented by the RBParameters object
Previously, we determined the number of steps stored on an
RBParameters object "dynamically" by simply checking how many steps
the user had stored for the first (alphabetically) parameter.

Unfortunately, this does not work well for a common use case, which is
when there are no parameters actually stored on the RBParameters
object, but it is still being used to represent multiple steps. In
this case, the number of steps which the RBParameters object
represents is returned by the n_steps() API.