Error in EigenSystem::solve if we have constraints
At least with SLEPc's eigen solver and the assembly routines in
MOOSE, the result of a solve does not honor constraints such as hanging
nodes. We do have `CondensedEigenSystem` which is a class intended
for use with constraints. Unrelated to this PR that class will need to be
updated in order to support shell matrices if MOOSE is to be able to
use it