Add MeshRefinement::_allow_unrefined_patches flag and accessor
This flag defaults to false in order to maintain the original behavior
of the code, which was to always eliminate unrefined element
patches. If you set it to true, then you can allow such patches to
persist, which may be useful in e.g. 1D meshes where having unrefined
patches does not introduce additional hanging nodes.