Use CFLAGS from autoconf-submodule
We weren't getting -O0 -g in contrib .c files in METHOD=dbg before,
which was a bit annoying when trying to debug.  Trying to debug is still
a bit annoying (there's some m4 code generation going on in NetCDF that
seems to be confusing my gdb about line numbers) but it's a lot better
than it was.