Handle overlapping global indices in Metis
In general this might not be a safe thing to do (disambiguating via id
rather than unique_id is subject to possible reordering that could
scramble restarts), but doing it in this particular case where we're
just trying to give Metis a good initial iterate should be robust.

This fixes a parallel failure in testExodusIGASidesets (with a
--disable-unique-id build) for me; the failure could crop up with any
mesh with perfectly-aligned overlapping elements, though.