Only test for simplexes
What's fascinating is that the two different assembly options
give identical results with simplexes but for tensor product elements
the "finite element" implementation gives a higher error. Moreover,
even for the lower error "algebraic" implementation the convergence
rates are not sharply k + 1, k + 1, and k + 2 for q, u, and u_enriched
respectively for the tensor product elements like they are for the
simplexes. My hypothesis for these observations is that the bilinearity
of the tensor product elements messes things up