Just always do the increment
There is a fundamental problem with this. The problem is that the
computed increment for the mixed system is nonzero even if the
increment is zero for the LM system. So basically I think trying
to increment the mixed solution outside of a global incrementing is
just not gonna work