*Really* kinda support .pvtu reads
This is a serial read of a parallel file, which is embarrassing, but
it's working, and without better documentation or unit testing I don't
want to push it harder than this until we have a user who needs it.

Fun fact: right now when I google "vtkxmlpunstructuredgridreader mpi",
the top result is the ticket "(Lack of) Documentation and Examples of
Parallel (MPI) VTK".  Answers there were helpful, but not enough.