Clarify valid_params function and add support for multiple samples
Previously we iterated over the serialized params iterator, which returns the name/value
for each parameter and value, including across steps. However, instead of using that
value, we called the params.get_value() function, which in all cases returns the 0th
step value. So steps/samples would silently fail to check anything but the first value.

Now we explicitly check all the step values if they're in the defined range,
and we give more clear messages if that's not the case.