Replace NumericVector with std::vector in RBConstructionBase
NumericVector was being used to determine which sample indices are local vs global.
But since the indices are the same across all parameters, and we don't need to do
any synchronization, it's overkill.

Note that previously, the NumericVector was sized to the global size, but this is wasteful
with std::vector. So instead, the std::vector's are sized only for the local number of
samples, and the global indices are adjusted when reading.

Removing NumericVector also opens up the possibility for us to store more complicated
datatypes in the future, e.g. a std::vector<Number> instead of a Number.