[bugfix] handle empty parameters case in get_params_from_training_set
Previously the libmesh_assert here would trip (in DBG mode) if training_parameters was empty.
When I switched it to a libmesh_error_msg_if (which also acts in OPT), it errored on "working"
tests which had empty training_parameters. Previously these tests passed in OPT mode because
the effective behavior was to ignore the assert and return an empty RBParameters.

Now I handle the empty parameters case explicitly by returning the empty RBParameters,
which is what the caller is expecting in this case. The index checking is only performed if
some parameters exist.