Don't duplicate mapping code for xyz
AFAICT there's not really a strong reason for the XYZ map class
to have it's own implementation for `compute_face_map` other than
to error for dim=1. I could be wrong though. This refs segmentation
faults on idaholab/moose#26456 (see https://civet.inl.gov/job/2068980/
for example) which were occurring because the XYZ `compute_face_map`
implementation was not doing checking of `if (calculate_foo)`