Always number IGA nodes consecutively
If we don't do so, then we can easily trigger renumbering that other
code paths are not prepared for.

On a DistributedMesh, this means numbering new nodes manually, to
override the default "stay prepared to mix replicated with distributed
node addition" sparse numbering.

This is approximately one-third of the *right* fix for this problem, but
it's a *sufficient* fix; I have a set of tests in a MOOSE branch which
(when run with --distributed-mesh) fail on master and pass on this
branch.