No Dirichlet interp on non-Lagrange non-vertices
In the best case we notice that the non-vertex hosts multiple DoFs and
we fail an assertion; in the worst case just setting a coefficient value
to a function value would give garbage here.

Never noticed this with libMesh/GRINS/etc codes, but MOOSE trips it as
soon as I try adding a DirichletBoundary there, since MOOSE generates
nodesets from sidesets.