Don't do nontrivial calculations just in DEBUG for Slepc solver
This removes the printing of eigenvalues and associated computation
of errors from the `solve` method for `SlepcEigenSolver`, which
was previously only occurring for `dbg` mode. Instead I've added
a `print_eigenvalues` routine that can be used to print eigenvalues,
not attached to any given compilation method. This code necessitated
assembly of the B matrix for the inverse power method in MOOSE when it is
not strictly needed