Entities BoundingBox testing in GmshIO
Right now we ignore the point locations and the curve/surface/volume
bounding box extents in a GMSH $Entities section; this led to unexpected
behavior in #3899

Let's insist that those locations and bounding boxes are accurate.
False positives for these tests are an easy fix (edit the $Entities),
whereas missed true positives lead to erroneous simulations.