Add SCALED_JACOBIAN ElemQuality enumeration
This enumeration is a straightforward extension of the JACOBIAN
enumeration, however it requires a bit more careful handling since it
involves dividing by edge lengths, which may be zero.

Add the new enumeration at the end of the list since the last time
new enumerations were added in the middle of the list, they broke
some code that was relying on hard-coded values for the enumerations.