Add Elem::qual_bounds() support for SCALED_JACOBIAN metric
According to the Cubit documentation [0-3], the JACOBIAN metric doesn't
really have an "acceptable range", presumably because it can take on
any value between -inf and inf (depending on the size of the element)
while the element can still have "good" quality.  So, the quality
bounds that we were previously using for the JACOBIAN metric should
actually be used for the SCALED_JACOBIAN metric instead. In fact, in
newer versions of the documentation [4], JACOBIAN is not even listed
as a potential metric, probably because it doesn't make much sense to
use it without some kind of scaling.

[0]: https://cubit.sandia.gov/files/cubit/16.08/help_manual/WebHelp/mesh_generation/mesh_quality_assessment/mesh_quality_assessment.htm
[1]: https://cubit.sandia.gov/files/cubit/16.08/help_manual/WebHelp/mesh_generation/mesh_quality_assessment/tetrahedral_metrics.htm
[2]: https://cubit.sandia.gov/files/cubit/16.08/help_manual/WebHelp/mesh_generation/mesh_quality_assessment/hexahedral_metrics.htm
[3]: https://cubit.sandia.gov/files/cubit/16.08/help_manual/WebHelp/mesh_generation/mesh_quality_assessment/quadrilateral_metrics.htm
[4]: https://coreform.com/cubit_help/mesh_generation/mesh_quality_assessment/triangular_metrics.htm