Allow for node permutations in Netgen test
I can't iterate through all 16! combinations but in the first 10K
or so we were getting a bit under 90% successes vs 10% failures.

This probably takes the lifetime of the universe to complete so we'll
leave it off by default, but it might be good for regression hunting in
the future so we'll leave it in.