Add --verbose option to unit tests driver
When we get an error message (e.g. from -fsanitize) that doesn't
actually cause the unit test to fail, or when we get a failure in
parallel that can't be recovered from, there was previously no way to
figure out where the error occurred except by bisecting names with
`--re`.  Simply seeing the test names printed out is much friendlier
when debugging, albeit too wordy to make a default behavior.