Double quote $LIBMESH_DIR in run.sh scripts
As suggested by shellcheck.net, double quoting prevents "globbing" and
"word splitting". So, doing this should help with cases where the
$LIBMESH_DIR path has spaces in it.

Refs #3988

This commit was cherry-picked from a later version of the master
branch that has more examples than the 1.7.x release branch.  Thus,
the following files showed up as "conflict" during the cherry-pick.
We "fixed" the conflicts by simply deleting the changes to non-existent
files.

Conflicts:
	examples/eigenproblems/eigenproblems_ex4/run.sh
	examples/miscellaneous/miscellaneous_ex16/run.sh
	examples/vector_fe/vector_fe_ex6/run.sh
	examples/vector_fe/vector_fe_ex7/run.sh
	examples/vector_fe/vector_fe_ex8/run.sh
	examples/vector_fe/vector_fe_ex9/run.sh