Switch from /bin/bash -> /bin/sh in run.sh scripts
As discussed on #3988, this should be the more portable option both
because bash is not always in /bin on all systems, and because we
should not require bash to run a simple script when any
POSIX-compatible shell should be sufficient.

Like the previous cherry-picked patch, this one had "conflicts" with
files that did not exist on the 1.7.x release branch.

Conflicts:
	examples/eigenproblems/eigenproblems_ex4/run.sh
	examples/miscellaneous/miscellaneous_ex16/run.sh
	examples/vector_fe/vector_fe_ex6/run.sh
	examples/vector_fe/vector_fe_ex7/run.sh
	examples/vector_fe/vector_fe_ex8/run.sh
	examples/vector_fe/vector_fe_ex9/run.sh