Setup PETSc aux data in Petsc(Non)LinearSolver
Provision beyond hypre preconditioners, but simply create
empty setup function if LIBMESH_HAVE_PETSC_HYPRE is not
defined - so it's clear, both to us and the compiler, we
currently do nothing in those cases.

set_petsc_aux_data() is called in PetscNonlinearSolver::solve().
This is due to commit c10f89f in idaholab/moose#29072
which changed MOOSE to call initPetscOutputAndSomeSolverSettings()
*before* petscSetOptions(). The former calls init() on
PetscNonlinearSolver() which is where we used to call
set_petsc_aux_data(), which in turn calls PCSetFromOptions(),
at which point the options need to be set.