Fix compilation errors in reduced_basis_ex4
RBEIMEvaluation::project_qp_data_map_onto_system() was removed in
5fd32f3e, but reduced_basis_ex4, an example that used it, was never
updated accordingly.

This example requires additional updates in order for it have the
same behavior as it did previously. Namely, the
SimpleEIMEvaluation::project_qp_data_vector_onto_system()
override needs to be implemented, since at the moment it is
currently a no-op.