Drop code used to plot parametrized function and basis function in reduced_basis_ex4
It would be non-trivial to implement a working override of
SimpleEIMEvaluation::project_qp_data_vector_onto_system()
in the context of this example, and we have test coverage for this in
our internal code.  Since it was only being used to visualize
ancillary data and not the solution itself, this seems like a
reasonable compromise.