configure.ac: bump version to 1.9.0-pre
The autoconf version on master is not really important, but this is
what we typically do on master once a release (or release candidate)
has been created for the previous series. I suppose it could be useful
in the context of a non-release tarball (i.e. when there is no git repo
present to tell you the actual hash) but that seems like a very limited
use scenario.