Optimize SparseMatrix::read_matlab()
There's a bit of refactoring here from when I was trying to log
performance more cleanly to diagnose the problem, and then there's a
reordering of regex tests that speeds up the code 500% for me.

It turns out that the std::regex library is really slow, and is going to
stay slow until they decide to break ABI compatibility.  I don't want to
rewrite those entry tests without regex and I don't want to bring in
Google's RE library or even boost::regex, so this is the best we can do
for now.