Remove non-const-reference-returning NumericVector::type() API
The user should not be allowed to change the ParallelType of a
NumericVector on the fly. Ideally, this flag should be set prior to
the vector being initialized, and then not changed. If we do need to
support changing it, it would be better to have a dedicated
NumericVector::set_type() API that does this, but we would still need
to restrict when such an API would be allowed to be called.