Add `operator[]` for `DenseVector`
I think of `operator[]` as being most appropriate for already
allocated data whereas I usually associate `operator()` with
on-the-fly functor evaluations. Obviously we are not going to
remote `operator()` (and I believe some would make a math notation
argument for keeping it anyway which I'm not against), but I think
we should support `operator[]`. This allows some templated code that takes
things like `std::vector` and `std::array` to also work with `DenseVector`