QBase::init(ElemType) simple_type_only param
We can't deprecate this overload entirely, because there are use cases
where we e.g. want a 1D quadrature rule without wanting to create an
actual Edge2 to pass to the const Elem& overload, but we don't want
application codes to manually init() a quadrature rule in a way that
will break on general polygons and polyhedra.  So we'll provide a way to
say "no, we're not screwing up, we know we're specifying an ElemType
for which you don't need a full Elem".