Elem::type_to_n_*_map[POLYGON1] handling
Obviously we can't rely on static maps here in general anymore, since
we're not going to have a different ElemType and map entry for every
possible polyhedron.

So, add some invalid_uint entries for the general types, and handle
invalid_uint where we might run into it, and query the Elem itself
instead of these maps whereever we can.