Fix deprecation warnings in custom iterators
The master branch version of this commit:

commit e5b6f482eda42c11197a8f60f2a817e6f8a88830
Author: Casey Icenhour <cticenho@ncsu.edu>
Date:   Thu Mar 17 11:40:07 2022 -0600

    Fixes up deprecation warnings in custom iterators due to C++17

had conflicts with the 1.7.x branch due to other changes that were
made to variant_filter_iterator.h in the meantime, so instead of
cherry-picking that commit directly, I just recreated the relevant
bits of it. This fixes an annoying warning that was otherwise seen
over and over again when compiling on newer compilers.